cli: move the URL parsing code to a standalone package
This is to prepare its use in the cockroach-sql standalone program.

Release note: None
