+++ pkg/cli/BUILD.bazel
@@ -100,6 +100,7 @@ go_library(
         "//pkg/ccl/sqlproxyccl",
         "//pkg/ccl/sqlproxyccl/tenantdirsvr",
         "//pkg/cli/clicfg",
+        "//pkg/cli/clienturl",
         "//pkg/cli/clierror",
         "//pkg/cli/clierrorplus",
         "//pkg/cli/cliflags",
