+++ pkg/cli/clienturl/BUILD.bazel
@@ -1,0 +1,21 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "clienturl",
+    srcs = [
+        "client_url.go",
+        "doc.go",
+        "flags.go",
+    ],
+    importpath = "github.com/cockroachdb/cockroach/pkg/cli/clienturl",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/cli/cliflags",
+        "//pkg/cli/clisqlshell",
+        "//pkg/security/clientsecopts",
+        "//pkg/server/pgurl",
+        "@com_github_cockroachdb_errors//:errors",
+        "@com_github_spf13_cobra//:cobra",
+        "@com_github_spf13_pflag//:pflag",
+    ],
+)
